+++ src/domain/prices.js
@@ -3,8 +3,19 @@ import { gql } from "@apollo/client";
 import useSWR from "swr";
 import { ethers } from "ethers";
 
-import { USD_DECIMALS, CHART_PERIODS, formatAmount, sleep } from "../lib/legacy";
+import {
+  USD_DECIMALS,
+  CHART_PERIODS,
+  formatAmount,
+  sleep,
+  PRECISION,
+  getTokenInfo,
+  getFeeBasisPoints,
+  MINT_BURN_FEE_BASIS_POINTS,
+  TAX_BASIS_POINTS,
+} from "../lib/legacy";
 import { chainlinkClient } from "./common";
+import { getVisibleTokens } from "../config/Tokens";
 
 const BigNumber = ethers.BigNumber;
 
@@ -287,4 +298,37 @@ function getStablePriceData(period) {
     });
   }
   return priceData;
+}
+
+export function getLowestFeeTokenForBuyGlp(
+  chainId,
+  toAmount,
+  glpPrice,
+  usdgSupply,
+  totalTokenWeights,
+  infoTokens,
+  fromTokenAddress
+) {
+  if (!chainId || !toAmount || !infoTokens || !glpPrice || !usdgSupply || !totalTokenWeights) {
+    return;
+  }
+  const tokens = getVisibleTokens(chainId);
+
+  const usdgAmount = toAmount.mul(glpPrice).div(PRECISION);
+  return tokens
+    .map((token) => {
+      const fromToken = getTokenInfo(infoTokens, token.address);
+      const fees = getFeeBasisPoints(
+        fromToken,
+        usdgAmount,
+        MINT_BURN_FEE_BASIS_POINTS,
+        TAX_BASIS_POINTS,
+        true,
+        usdgSupply,
+        totalTokenWeights
+      );
+      return { ...token, fees };
+    })
+    .filter((t) => t.address !== fromTokenAddress && t.fees)
+    .sort((a, b) => a.fees - b.fees)[0];
 }
